[🚚move] 일정관리 유틸 경로 변경
