+++ src/utils/schedules/getDate.tsx
@@ -20,42 +20,3 @@ export const getLastDateOfMonth = (selectedDate: string): { formattedFirstDate: 
 
   return { formattedFirstDate, formattedLastDate };
 };
-
-interface DateDetails {
-  date: number;
-  month: number;
-  year: number;
-  currentMonth: boolean;
-  dayOfWeek: number;
-}
-
-/**
- * 2023-00-00 > object 반환
- * @param selectedDate
- * @returns
- */
-export const getDateDetails = (selectedDate: string): DateDetails => {
-  const selected = new Date(selectedDate);
-  const curMonth = new Date().getMonth() + 1;
-
-  return {
-    date: selected.getDate(),
-    month: selected.getMonth() + 1,
-    year: selected.getFullYear(),
-    currentMonth: selected.getMonth() + 1 === curMonth,
-    dayOfWeek: selected.getDay(),
-  };
-};
-
-/**
- * object > 2023-00-00 반환
- * @param dateDetails
- * @returns
- */
-export const formatDateString = (dateDetails: DateDetails) => {
-  const date = dateDetails.date;
-  const month = dateDetails.month;
-  const year = dateDetails.year;
-
-  return `${year}-${String(month).padStart(2, '0')}-${String(date).padStart(2, '0')}`;
-};
